Tools: added rp_remote_acquire
the continuous acquisition tool rp_remote_acquire has been adapted
to be compatible with the unmodified ecosystem releases 0.96 and
later. to build it call make in Test/rp_remote_acquire. it needs
the included bitstream ddrdump.bit to function, there is a command
line switch to do that automatically before acquisition starts.