Do not hang if delimiter is missing
If the text string returned from the SCPI server is not terminated with
self.delimiter (a CR/LF) then the rx_txt() function will hang, as it
only terminates when it sees the delimiter.

If the chunk returned is smaller than chunksize then we know there is no
more data to be read. We are safe to return at that point.

Stop adding the length of the delimiter to chunksize. For efficiency
chunksize needs to be a power of two and adding the delimiter made it no
longer a power of two. Also, there is no need to add space for the
delimiter. If the delimiter does not happen to fit in the current chunk
then it will be read in a subsequent read.

When checking whether the delimiter has been seen, stop looking at
chunk. Instead, look at msg. It could be the case that the message
length (without the delimiter) was one smaller than the buffer length.
This means that the delimiter would be split over two chunks and testing
chunk to see if it was terminated would fail. Test the msg instead.