Truncate msg correctly, don't exit early
msg needs to have the last two characters trimmed if they are `\r\n`.
The existing code was *returning* the last two characters.

Sometimes the socket is read before the remote buffer is full. In that
case the returned text may be shorter than the supplied read buffer and
they transmission may not be complete. Don't exit.