Fix type, broken links, formatting
Fix a typo.

Fix the formatting so the numbered headers do not run on into the paragraph.

Fix broken links. I had some trouble with the teaching link (last one on the page). I could not find any teaching references on redpitaya.readthedocs.io, so I chose to link to the Red Pitaya corporate site.